* temp commit

+++ worker.py
@@ -14,46 +14,57 @@ logger = logging.getLogger('worker')
 
 class Worker(object):
 
-    def __init__(self, name, funcfile, host='localhost', size=100, port=11300):
+    def __init__(self, name, cfgs, host='localhost', port=11300):
         self.queue = beanstalkc.Connection(host=host, port=port)
-        self.queue.use('%s' % name)
-        self.queue.watch('%s' % name)
+        self.queue.watch(name)
+        self.queue.use(name)
         self.name = name
-        self.cfgs = loadcfg(funcfile)
-        self.pool = gevent.pool.Pool(size)
+        self.cfgs = cfgs
 
     def run(self):
         while True:
-            job = self.queue.reserve()
-            self.pool.spawn(self.async_run, job)
+            job = self.queue.reserve(timeout=1)
+            if job is None: return
 
-    def async_run(self, job):
-        req = job.body
-        logger.debug('get: ' + req)
-        for m, n, p in dispatch(self.cfgs, req):
-            logger.info('%s run.' % p['handler'])
-            if m is None: m = []
-            rslt, reqs = p['function'](req, *m)
-            # howto process rslt
-            logger.info('result: %s' % str(rslt))
-            if reqs:
-                for req in reqs: self.request(req)
-        job.delete()
+            req = job.body
+            logger.debug('get: ' + req)
+            for m, n, p in dispatch(self.cfgs, req):
+                logger.info('%s run.' % p['handler'])
+                if m is None: m = []
+                rslt = p['function'](self, req, *m)
+                # howto process rslt
+                logger.info('result: %s' % str(rslt))
+            job.delete()
 
     def request(self, url, headers=None, body=None, method='GET'):
         self.queue.put(url)
         logger.debug('put: ' + str(url))
 
+def workgroup(name, funcfile, size=100, host='localhost', port=11300):
+    cfgs = loadcfg(funcfile)
+    pool = gevent.pool.Pool(size)
+    for n in xrange(size): pool.spawn(Worker(name, cfgs).run)
+    pool.join()
+
+def load_handler(handler):
+    cfgfile = handler['handler']
+    with open(cfgfile) as fi: cfg = yaml.load(fi.read())
+    for p in cfg['patterns']:
+        
+
 def loadcfg(cfgfile):
-    with open(cfgfile) as fi: cfgs = yaml.load(fi.read())
+    with open(cfgfile) as fi: cfg = yaml.load(fi.read())
     sys.path.append(path.dirname(cfgfile))
-    for cfg in cfgs:
-        cfg['base'] = urlparse(cfg['host'])
-        for p in cfg['patterns']:
-            modname, funcname = p['handler'].split(':')
-            if not modname: modname = cfg['file']
-            p['function'] = getattr(__import__(modname), funcname)
-            p['re'] = re.compile(p['match'])
+    cfg['handlers'] = [load_handler(h) for h in cfg['handlers']]
+
+    # for cfg in cfgs:
+    #     cfg['base'] = urlparse(cfg['host'])
+    #     for p in cfg['patterns']:
+    #         modname, funcname = p['handler'].split(':')
+    #         if not modname: modname = cfg['file']
+    #         p['function'] = getattr(__import__(modname), funcname)
+    #         p['re'] = re.compile(p['match'])
+
     sys.path.pop(-1)
     return cfgs
 
@@ -72,8 +83,8 @@ def dispatch(cfgs, url):
 def main():
     cfgs = loadcfg(sys.argv[1])
     pprint.pprint(cfgs)
-    if len(sys.argv) > 2:
-        for m, n, p in dispatch(cfgs, sys.argv[2]):
-            print m, n, p
+    # if len(sys.argv) > 2:
+    #     for m, n, p in dispatch(cfgs, sys.argv[2]):
+    #         print m, n, p
 
 if __name__ == '__main__': main()
